refactor: rebuild registry using composition
