+++ packages/catalog/src/registry/modules/packages.ts
@@ -1,0 +1,232 @@
+import {
+  type ModuleDefinition,
+  ModuleId,
+  TargetIdentity,
+  TargetKind,
+} from "@repo/domain/Catalog";
+import {
+  aiAgenticLoopContents,
+  aiChatServiceContents,
+  aiIndexContents,
+  aiLanguageModelContents,
+  aiMailboxEventsContents,
+  aiSampleToolkitContents,
+} from "../content/ai";
+import {
+  presenceClientGeneratorContents,
+  presenceIndexContents,
+  presenceServiceContents,
+} from "../content/presence";
+
+/**
+ * Package modules - infrastructure packages (ai, presence)
+ */
+export const packageModules: ReadonlyArray<typeof ModuleDefinition.Type> = [
+  {
+    id: ModuleId.make("ai"),
+    title: "AI Package",
+    description:
+      "Anthropic language model configuration and workflow utilities",
+    supportedOn: [
+      {
+        _tag: "identity",
+        identity: new TargetIdentity({
+          kind: TargetKind.make("package"),
+          name: "ai",
+        }),
+      },
+    ],
+    dependencies: [],
+    contributions: [
+      {
+        _tag: "file",
+        path: "{{targetPath}}/src/index.ts",
+        contents: aiIndexContents,
+      },
+      {
+        _tag: "file",
+        path: "{{targetPath}}/src/LanguageModel.ts",
+        contents: aiLanguageModelContents,
+      },
+      {
+        _tag: "file",
+        path: "{{targetPath}}/src/workflow/MailboxEvents.ts",
+        contents: aiMailboxEventsContents,
+      },
+      {
+        _tag: "pkg-json-entry",
+        path: "{{targetPath}}/package.json",
+        field: "dependencies",
+        name: "@effect/ai-anthropic",
+        value: "4.0.0-beta.59",
+      },
+      {
+        _tag: "pkg-json-entry",
+        path: "{{targetPath}}/package.json",
+        field: "dependencies",
+        name: "effect",
+        value: "4.0.0-beta.59",
+      },
+    ],
+  },
+  {
+    id: ModuleId.make("ai-sample-toolkit"),
+    title: "Sample Toolkit",
+    description: "Sample AI toolkit with calculator, echo, and time tools",
+    supportedOn: [
+      {
+        _tag: "identity",
+        identity: new TargetIdentity({
+          kind: TargetKind.make("package"),
+          name: "ai",
+        }),
+      },
+    ],
+    dependencies: [],
+    contributions: [
+      {
+        _tag: "file",
+        path: "{{targetPath}}/src/toolkits/SampleToolkit.ts",
+        contents: aiSampleToolkitContents,
+      },
+      {
+        _tag: "barrel-export",
+        barrelPath: "{{targetPath}}/src/index.ts",
+        exportPath: "./toolkits/SampleToolkit",
+      },
+    ],
+  },
+  {
+    id: ModuleId.make("ai-chat-service"),
+    title: "Chat Service",
+    description:
+      "AI chat service with agentic loop for streaming tool-augmented conversations",
+    supportedOn: [
+      {
+        _tag: "identity",
+        identity: new TargetIdentity({
+          kind: TargetKind.make("package"),
+          name: "ai",
+        }),
+      },
+    ],
+    dependencies: [
+      {
+        requiredTarget: {
+          identity: new TargetIdentity({
+            kind: TargetKind.make("package"),
+            name: "domain",
+          }),
+        },
+        requiredModule: {
+          target: new TargetIdentity({
+            kind: TargetKind.make("package"),
+            name: "domain",
+          }),
+          moduleId: ModuleId.make("domain-chat"),
+        },
+      },
+      {
+        requiredModule: {
+          target: new TargetIdentity({
+            kind: TargetKind.make("package"),
+            name: "ai",
+          }),
+          moduleId: ModuleId.make("ai-sample-toolkit"),
+        },
+      },
+    ],
+    contributions: [
+      {
+        _tag: "file",
+        path: "{{targetPath}}/src/services/ChatService.ts",
+        contents: aiChatServiceContents,
+      },
+      {
+        _tag: "file",
+        path: "{{targetPath}}/src/workflow/AgenticLoop.ts",
+        contents: aiAgenticLoopContents,
+      },
+      {
+        _tag: "pkg-json-entry",
+        path: "{{targetPath}}/package.json",
+        field: "dependencies",
+        name: "@repo/domain",
+        value: "workspace:*",
+      },
+      {
+        _tag: "barrel-export",
+        barrelPath: "{{targetPath}}/src/index.ts",
+        exportPath: "./services/ChatService",
+      },
+      {
+        _tag: "barrel-export",
+        barrelPath: "{{targetPath}}/src/index.ts",
+        exportPath: "./workflow/AgenticLoop",
+      },
+    ],
+  },
+  {
+    id: ModuleId.make("presence"),
+    title: "Presence Package",
+    description:
+      "Real-time presence tracking service with PubSub and client generation",
+    supportedOn: [
+      {
+        _tag: "identity",
+        identity: new TargetIdentity({
+          kind: TargetKind.make("package"),
+          name: "presence",
+        }),
+      },
+    ],
+    dependencies: [
+      {
+        requiredTarget: {
+          identity: new TargetIdentity({
+            kind: TargetKind.make("package"),
+            name: "domain",
+          }),
+        },
+        requiredModule: {
+          target: new TargetIdentity({
+            kind: TargetKind.make("package"),
+            name: "domain",
+          }),
+          moduleId: ModuleId.make("domain-websocket"),
+        },
+      },
+    ],
+    contributions: [
+      {
+        _tag: "file",
+        path: "{{targetPath}}/src/index.ts",
+        contents: presenceIndexContents,
+      },
+      {
+        _tag: "file",
+        path: "{{targetPath}}/src/services/ClientGenerator.ts",
+        contents: presenceClientGeneratorContents,
+      },
+      {
+        _tag: "file",
+        path: "{{targetPath}}/src/services/PresenceService.ts",
+        contents: presenceServiceContents,
+      },
+      {
+        _tag: "pkg-json-entry",
+        path: "{{targetPath}}/package.json",
+        field: "dependencies",
+        name: "@repo/domain",
+        value: "workspace:*",
+      },
+      {
+        _tag: "pkg-json-entry",
+        path: "{{targetPath}}/package.json",
+        field: "dependencies",
+        name: "effect",
+        value: "4.0.0-beta.59",
+      },
+    ],
+  },
+];
